+++ frontend/src/components/Navbar.jsx
@@ -1,4 +1,4 @@
-import logo from "./assets/logo.svg";
+import logo from "../assets/logo.svg";
 import { Link } from "react-router-dom";
 
 export function Navbar() {
