+++ frontend/src/components/Sort.jsx
@@ -1,6 +1,6 @@
-import { useReduce } from "./Reducer-context";
+import { useReduce } from "../reducer-context/Reducer-context";
 
-export default function Sort({ setshowsort }) {
+export function Sort({ setshowsort }) {
   let { dispatch, sortBy } = useReduce();
 
   return (
