+++ frontend/src/components/index.js
@@ -1,0 +1,4 @@
+export { Navbar } from "./Navbar";
+export { Navigator } from "./Navigator";
+export { Filter } from "./Filter";
+export { Sort } from "./Sort";
