+++ frontend/src/pages/Liked.jsx
@@ -1,16 +1,16 @@
-import { useReduce } from "../Reducer-context";
+import { useReduce } from "../reducer-context/Reducer-context";
 import { getSortedData, getFilteredData } from "../App";
-import Navigator from "../Navigator";
+import { Navigator, Navbar } from "../components";
 import { useEffect } from "react";
 import { Link } from "react-router-dom";
-import { Navbar } from "../Navbar";
+// import { Navbar } from "../Navbar";
 import { toast } from "react-toastify";
 import {
   history_video_add_call,
   liked_video_delete_call
 } from "../api/serverRequests";
 
-export default function Liked() {
+export function Liked() {
   useEffect(() => {
     dispatch({ type: "CLEAR_FILTER" });
   }, []);
