+++ frontend/src/pages/NotFound.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-export default function NotFound() {
+export function NotFound() {
   return (
     <div className="notfound">
       <div>
