+++ frontend/src/pages/index.js
@@ -1,0 +1,9 @@
+export { Home } from "./Home";
+export { History } from "./History";
+export { Liked } from "./Liked";
+export { Login } from "./Login";
+export { NotFound } from "./NotFound";
+export { Playlist } from "./Playlist";
+export { Signup } from "./Signup";
+export { User } from "./User";
+export { VideoBlock } from "./VideoBlock";
